+++ js/data.js
@@ -1,0 +1,40 @@
+const COUNT_PHOTO = 25;
+const COUNT_AVATARS = 6;
+const MIN_COUNT_COMMENTS = 0;
+const MAX_COUNT_COMMENTS = 30;
+const MIN_COUNT_LIKES = 15;
+const MAX_COUNT_LIKES = 200;
+const DESCRIPTION = [
+  'Застывший момент красоты.',
+  'Истории, рассказанные светом и тенью.',
+  'Каждый кадр – это новый взгляд.',
+  'Вдохновение в деталях.',
+  'Поймано мгновение, пойман смысл.',
+  'Когда изображение говорит больше тысячи слов.',
+  'Гармония в объективе.',
+  'Отражение мира вокруг нас.',
+  'Взгляд, полный эмоций.',
+  'Путешествие в одном снимке.'
+];
+const NAMES = [
+  'Александр',
+  'София',
+  'Максим',
+  'Анна',
+  'Дмитрий',
+  'Екатерина',
+  'Иван',
+  'Ольга',
+  'Михаил',
+  'Анастасия'
+];
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export {COUNT_PHOTO,DESCRIPTION, NAMES, MESSAGES, COUNT_AVATARS, MIN_COUNT_COMMENTS, MAX_COUNT_COMMENTS, MIN_COUNT_LIKES, MAX_COUNT_LIKES};
